pyproject.toml
[tool.poetry]
name = "privacypacking"
version = "0.1.0"
description = ""
authors = ["Pierre Tholoniat <pierre@tholoniat.com>"]

[tool.poetry.dependencies]
python =  ">=3.8,<3.10"
simpy = "^4.0.1"
loguru = "^0.5.3"
jupyter = "^1.0.0"
notebook = "^6.4.0"
jupyterlab = "^3.0.16"
numpy = "^1.19.0"
PuLP = "^2.4"
gurobipy = "^9.1.2"
scipy = "^1.7.0"
plotly = "^5.1.0"
kaleido = "==0.2.1"
pandas = "^1.3.0"
dash = "^1.21.0"
torch = "1.8.1"
opacus = "^0.14.0"
PyYAML = "^5.4.1"
desmod = "^0.6.1"
ray = {extras = ["tune"], version = "^1.12.1"}

tensorboard = "^2.6.0"
tensorflow = "^2.6.0"
mip = "^1.13.0"
typer = "^0.4.0"
"amazon.ion" = "^0.7.0"
scalene = "^1.3.16"
tqdm = "^4.62.3"
omegaconf = "^2.1.1"
mlflow = "^1.23.1"
snakeviz = "^2.1.1"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.9.2"
nb-clean = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
max-line-length = 79